refactor: moving SchemaVersion and RdfStoreData to graph-commons
